Add test before copying dusk.desktop due to changes in core utils which makes cp -n exit with error if the file exists, resulting in Make treating it as a build failure.

+++ Makefile
@@ -52,7 +52,7 @@ install: all
 	sed "s/VERSION/${VERSION}/g" < dusk.1 > ${DESTDIR}${MANPREFIX}/man1/dusk.1
 	chmod 644 ${DESTDIR}${MANPREFIX}/man1/dusk.1
 	mkdir -p /usr/share/xsessions
-	cp -n dusk.desktop /usr/share/xsessions/
+	test -f /usr/share/xsessions/dusk.desktop || cp -n dusk.desktop /usr/share/xsessions/
 	chmod 644 /usr/share/xsessions/dusk.desktop
 
 uninstall:
